LPS-109676 Move role type contributors to roles-admin-impl
