+++ modules/apps/roles/roles-admin-impl/src/main/java/com/liferay/roles/admin/internal/role/type/contributor/RegularRoleTypeContributor.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.roles.admin.web.internal.role.type.contributor;
+package com.liferay.roles.admin.internal.role.type.contributor;
 
 import com.liferay.portal.kernel.model.Role;
 import com.liferay.portal.kernel.model.role.RoleConstants;
